requirements.txt
avisdk==22.1.3
requests[SOCKS]==2.31.0
tabulate==0.9.0